move compareLabelFn() to @finos/legend-art
