+++ packages/legend-studio/src/components/editor/command-center/ProjectSearchCommand.tsx
@@ -16,11 +16,11 @@
 
 import { useState, useRef, useEffect } from 'react';
 import { observer } from 'mobx-react-lite';
-import { compareLabelFn } from '@finos/legend-shared';
 import { FaCaretDown } from 'react-icons/fa';
 import { MdMoreHoriz } from 'react-icons/md';
 import type { SelectComponent } from '@finos/legend-art';
 import {
+  compareLabelFn,
   DropdownMenu,
   NonBlockingDialog,
   createFilter,
